requirements.txt
pytest
fastapi
email-validator
python-dotenv
requests
pydantic
Faker~=22.5.1
uvicorn==0.34.0
fastapi-pagination
psycopg2-binary
sqlalchemy
sqlmodel